feat: filter notes by authenticated user

+++ app/Http/Controllers/NoteController.php
@@ -4,9 +4,6 @@ namespace App\Http\Controllers;
 
 use App\Models\Note;
 use Illuminate\Http\Request;
-use Illuminate\Database\Query\Builder;
-use Illuminate\Support\Facades\DB;
-use Illuminate\View\View;
 
 class NoteController extends Controller
 {
@@ -15,11 +12,10 @@ class NoteController extends Controller
      */
     public function index()
     {
-        $notes = Note::query()->orderBy('created_at', 'desc')->paginate();
-
-
-        // dd($notes); var_dump equivalent
-
+        $notes = Note::query()
+            ->where('user_id', request()->user()->id)
+            ->orderBy('created_at', 'desc')
+            ->paginate();
         return view('note.index', ['notes' => $notes]);
     }
 
@@ -40,10 +36,10 @@ class NoteController extends Controller
             'note' => ['required', 'string']
         ]);
 
-        $data['user_id'] = 1;
+        $data['user_id'] = $request->user()->id;
         $note = Note::create($data);
 
-        return to_route('note.show', $note)->with('message', 'Note was created');
+        return to_route('note.show', $note)->with('message', 'Note was create');
     }
 
     /**
@@ -51,6 +47,9 @@ class NoteController extends Controller
      */
     public function show(Note $note)
     {
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
         return view('note.show', ['note' => $note]);
     }
 
@@ -59,6 +58,9 @@ class NoteController extends Controller
      */
     public function edit(Note $note)
     {
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
         return view('note.edit', ['note' => $note]);
     }
 
@@ -67,6 +69,9 @@ class NoteController extends Controller
      */
     public function update(Request $request, Note $note)
     {
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
         $data = $request->validate([
             'note' => ['required', 'string']
         ]);
@@ -81,6 +86,9 @@ class NoteController extends Controller
      */
     public function destroy(Note $note)
     {
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
         $note->delete();
 
         return to_route('note.index')->with('message', 'Note was deleted');
